replaced set_world_docs with world_description property
