+++ src/Meta.py
@@ -3,10 +3,26 @@ from BaseClasses import Tutorial
 from worlds.AutoWorld import World, WebWorld
 from .Data import meta_table
 
-enable_region_diagram = meta_table.get("enable_region_diagram", False)
+##############
+# Meta Classes
+##############
+class ManualWeb(WebWorld):
+    tutorials = [Tutorial(
+        "Multiworld Setup Guide",
+        "A guide to setting up manual game integration for Archipelago multiworld games.",
+        "English",
+        "setup_en.md",
+        "setup/en",
+        ["Fuzzy"]
+    )]
 
-def set_world_doc(base_doc: str):
-    if meta_table.get("docs", {}).get("apworld_description", []):
+#################################################
+# Convert meta.json data to properties
+#################################################
+def set_world_description(base_doc: str) -> str:
+    if isinstance(meta_table["docs"]["apworld_description"], str):
+        base_doc = meta_table["docs"]["apworld_description"]
+    else:
         fullstring = ""
         for line in meta_table["docs"]["apworld_description"]:
             fullstring += "\n" + line
@@ -29,6 +45,7 @@ def set_world_webworld(web: WebWorld) -> WebWorld:
         if Web_Config.get("tutorials", []):
             tutorials = []
             for tutorial in Web_Config.get("tutorials", []):
+                # Converting json to Tutorials
                 tutorials.append(Tutorial(
                     tutorial.get("name", "Multiworld Setup Guide"),
                     tutorial.get("description", "A guide to setting up manual game integration for Archipelago multiworld games."),
@@ -39,3 +56,16 @@ def set_world_webworld(web: WebWorld) -> WebWorld:
                 ))
             web.tutorials = tutorials
     return web
+
+#################
+# Meta Properties
+#################
+world_description: str = set_world_description("""
+    Manual games allow you to set custom check locations and custom item names that will be rolled into a multiworld.
+    This allows any variety of game -- PC, console, board games, Microsoft Word memes... really anything -- to be part of a multiworld randomizer.
+    The key component to including these games is some level of manual restriction. Since the items are not actually withheld from the player,
+    the player must manually refrain from using these gathered items until the tracker shows that they have been acquired or sent.
+    """)
+world_webworld: ManualWeb = set_world_webworld(ManualWeb())
+
+enable_region_diagram = meta_table.get("enable_region_diagram", False)
